requirements.txt
asyncio~=3.4.3
pandas~=2.2.3
aiofiles~=24.1.0
python-dotenv~=1.0.1
colorlog~=6.8.2